Link Operator to foreground but not background
The Normal group sets both foreground and background.
When 'cursorline' is enabled and the current line contains an Operator,
the background defined for Normal "wins" and overrides the background
defined for CursorLine.

This change links Operator to GruvboxFg1 instead
which uses the same foreground but does not override the background.